Add animated center line to GridAlternating component
