+++ src/pages/{storyblokEntry.full_slug}.tsx
@@ -1,6 +1,5 @@
-import React, { useRef } from 'react';
+import React from 'react';
 import { graphql } from 'gatsby';
-import { useScroll, useSpring, m } from 'framer-motion';
 import { StoryblokComponent, storyblokEditable } from 'gatsby-source-storyblok';
 import { useStoryblokState } from '../hooks/useStoryblokState';
 import { Hero } from '../components/Hero/Hero';
@@ -14,14 +13,6 @@ const StoryblokEntry = ({ data }) => {
   story = useStoryblokState(story);
 
   const components = story.content.body?.map((blok) => (<StoryblokComponent blok={blok} key={blok._uid} />));
-
-  const gridRef = useRef(null);
-  const { scrollYProgress } = useScroll({ target: gridRef, offset: ['-300px', '200px'] });
-  const scaleY = useSpring(scrollYProgress, {
-    stiffness: 100,
-    damping: 30,
-    restDelta: 0.001,
-  });
 
   return (
     <Layout>
@@ -42,24 +33,21 @@ const StoryblokEntry = ({ data }) => {
             Test1 Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam non aliquam purus.
           </Paragraph>
         </GridAlternating>
+        <GridAlternating startOnRight>
+          {['test1', 'test2', 'test3']}
+        </GridAlternating>
       </Container>
       <Container bgColor="white" pt={9}>
-        <div className="su-relative su-mx-auto" ref={gridRef}>
-          <GridAlternating className="md:su-gap-x-[6rem] lg:su-gap-x-[10rem] xl:su-gap-x-[20rem] 2xl:su-gap-x-[28rem]">
-            <Paragraph font="serif" variant="overview" className="2xl:su-max-w-500 su-justify-self-end">
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam non aliquam purus.
-              Suspendisse efficitur sodales lorem. Nullam non aliquam purus.
-            </Paragraph>
-            <Paragraph font="serif" variant="overview" className="2xl:su-max-w-500 su-rs-pb-9">
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam non aliquam purus.
-              Suspendisse efficitur sodales lorem. Nullam non aliquam purus.
-            </Paragraph>
-          </GridAlternating>
-          <m.div
-            className="su-w-2 su-absolute su-top-0 su-left-[50%] su--ml-1 su-block su-bg-black su-h-full su-origin-top"
-            style={{ scaleY }}
-          />
-        </div>
+        <GridAlternating addCenterLine className="md:su-gap-x-[6rem] lg:su-gap-x-[10rem] xl:su-gap-x-[20rem] 2xl:su-gap-x-[28rem]">
+          <Paragraph font="serif" variant="overview" className="2xl:su-max-w-500 su-justify-self-end">
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam non aliquam purus.
+            Suspendisse efficitur sodales lorem. Nullam non aliquam purus.
+          </Paragraph>
+          <Paragraph font="serif" variant="overview" className="2xl:su-max-w-500 su-rs-pb-9">
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam non aliquam purus.
+            Suspendisse efficitur sodales lorem. Nullam non aliquam purus.
+          </Paragraph>
+        </GridAlternating>
       </Container>
       <div {...storyblokEditable(story.content)}>
         {components}
